bean-hwr/requirements.txt
click==6.7
DBUtils==1.2
django==1.11.4
Flask==0.12.2
gevent==1.2.2
greenlet==0.4.12
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
numpy==1.13.1
olefile==0.44

Pillow==4.2.1
psycopg2==2.7.3
pycrypto==2.6.1
pytz==2017.2
sqlalchemy==1.1.13
websocket-client==0.44.0
websocket-server==0.4
Werkzeug==0.12.2
pyyaml
web3
Crypto
web.py==0.40.dev0